refactor(WaitForHostElement): use new api
Also, change back the render() method from
returning with `return el && children;`.
Reason being, when el is not null and not
a React-Element React will think that
nothing is returned from the render.
Besides this there was a memory-leak warning
coming from React when I used this return
statement. As this not a critical change
I decided to go forward with the working
solution for now instead of debugging,
and get back to it when we are on track
with the Xmas campaign.