tests(PageCampaign): remove unnecessary test
Right now we don't really have any kind of logic
in this component, only wrapping the <ComponentBuilder />,
so a test to check if that is used wouldn't give us too much benefit.
(The test would have to be changed every time the code changes).
The rendering is checked by the Storybook.
Will add tests as soon as we have any kind of business logic.