refactor(PromotionOptInButton): Store player opt-in preference between page reloads
**Note**
Although a timestamp is being added when storing the preference, there's not yet a feature to clean up those keys after X amount of time.
This should be assessed after this PR so that we know how many tracking places will be used and for how long those promotions generally run for.