Allow using hooks directly inside a story (#483)
wanted to write stuff like this:

  stories.add("Default", () => {
    const inputProps = useInput();
    return <TextInput {...inputProps} />;
  });

but it was causing "Invalid hook call" which makes sense, "story" in
addDecorator is what you pass as second arg to stories.add. Hook was
called from regular function not react component.

This commit changes that behaviour - now 2nd arg is treated as new react
component.