change NODE_ENV back to test
It looks like it was done back when we had 2 env vars for tests:
"AUTOMATED_TESTS" & "TEST". Supposedly to distinguish between errors from test
env and CI on rollbar. We don't need that distinguishment anymore (in test env
we only log to console) and non-default env can lead to weird issues.

One example of why it's bad is this macro that we use for storyshots:
https://github.com/storybookjs/require-context.macro/blob/master/require-context.macro.js#L5-L14