feat(Models/Apollo): add mutations
Collecting the mutations in one place so they are easier to reference in components.
Right now we only have mutations for Sports, but it can (and should) change in the future.

Bare in mind that this is not new code, but something that is moved over from the sports folder.