feat(Models/Apollo): add common queries
These queries can be referenced later by the components.
This is probably not the way we would like to take in the future,
we would rather like to co-locate the queries at the components
(as it is up to every component to decide what data it needs).

Although currently it is something that Sports is depending on, and
it wouldn't make sense to refactor it now. Probably we will get back
to this later.