Fix GamesVirtualList not updating properly
Some time ago I added `listHash` property, to trigger Virtual List updates when
length of the list isn't changing but elements change order. I though that it
will solve the issues we had at that point. Now, that this list is used in more
places we started to see additional problems.

It seems that the only way to solve issues with updates not coming through is
to foce developers to change `listHash` every time update is required (even if
list is changing completly).