Remove reel race hooks from hooks/index.js
When test are running any import from `Utils/hooks` resolves all exports that
are there. Latest addition of game activity aware thingis caused all tests that
were using this path to break.

useGameActivityAwareValue is using context which needs to be present and valid
for it to work. Resolving this file was throwing bunch of errors.
There is no simple way to automock this, new file inside __mocks__ will only be
used by storybook (because of our custom config).