Hide temporarly blue ribbon win animation
Currently we show the win animation if the player is in a game, and it is
assumed that the win animation is related to the game they are playing.
However with Blue Ribbon Sports implementation it is possible for a jackpot to
drop which is unrelated to the slot game currently being played.

We will add proper fix for this as part of cms configuration for
multiple jackpots.