Send source maps mapping properly
Previously I was sending `./bundleName.hash.js` and I though rollbar is smart
enough to figure out that this file should be used when it gets requested ie.:
  https://www.casumo.com/react-stack/js/bundleName.hash.js
unfortunatley it's not the case. We need to send whole url without "https:".
Because of that each file needs to be send twice (once for casumo.com and once
for casumo.es), from what I see requests are always for www.casumo.*,
but we might need to add two additional urls (without www) if we see broken
source maps.