Fix pots not updating on game page
This bug was introduced in #1668, previously BlueRibbonChristmasCampaignNotifi-
cationsContainer was rendered on game page and useBlueRibbonAutoOptIn was
called inside of it. This hook now has a little bit misleading name - it just
starts the feed, which is required to get pot updates - which will be addressed
in refactor after campaign launch.