Add handling for SG digital games
PDF from 12/2020 says that this is only available on GCM enabled SG content and
they are in the process of rolling this feature out across all integrated Game
Providers with a completion date targeted for the end of 03/2021.
Ofcourse it was bullshit and they didn't delivered it yet.
If you need to test something rudolph-gone-wild is working currently.