Fix Microgaming integration
- gameActivityStatusSource was not set properly and because of it events were
  not send.

There are games that work nicely ie. cabaret-royale, there are games that work
partially ie. roman-power (where you don't get event when auto spins start and
our integration doesn't know that game is currently busy) and there are games
that don't work at all, like 9-masks-of-fire which - obviously - is the only
game mentioned in this comment that is used in blue ribbon campaign. 9 masks
sends only one event - playerInformationComplete which is part of the setup